agent-blueprint/agentcore-runtime-a2a-stack/research-agent/requirements.txt
# Research Agent Dependencies
# Using same versions as chatbot-app/agentcore for compatibility

# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Strands Agent Framework
# Using 1.14.0 for compatibility
strands-agents==1.14.0
strands-agents-tools==0.2.3

# AWS & Bedrock
boto3>=1.40.1
botocore>=1.34.0

# HTTP
httpx>=0.24.0
anyio>=3.6.0

# Web Search
ddgs>=9.0.0

# Wikipedia
wikipedia-api>=0.6.0

# HTML Parsing
beautifulsoup4>=4.12.0

# Document Generation
python-docx

# AgentCore for Gateway integration
bedrock-agentcore

# A2A SDK for agent-to-agent communication
a2a-sdk>=0.3.20

# OpenTelemetry (for observability)
aws-opentelemetry-distro>=0.10.1
